services: restructure base AI services into modules
