+++ src/pipecat/observers/loggers/llm_log_observer.py
@@ -18,7 +18,7 @@ from pipecat.frames.frames import (
 from pipecat.observers.base_observer import BaseObserver
 from pipecat.processors.aggregators.openai_llm_context import OpenAILLMContextFrame
 from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
-from pipecat.services.ai_services import LLMService
+from pipecat.services.llm_service import LLMService
 
 
 class LLMLogObserver(BaseObserver):
