+++ src/pipecat/services/__init__.py
@@ -37,4 +37,4 @@ class DeprecatedModuleProxy:
     def __getattr__(self, attr):
         if attr in self._globals:
             return _warn_deprecated_access(self._globals, attr, self._old, self._new)
-        raise AttributeError(f"module 'pipecat.{self._old}' has no attribute '{attr}'")
+        raise AttributeError(f"module 'pipecat.services.{self._old}' has no attribute '{attr}'")
