+++ src/pipecat/services/anthropic/llm.py
@@ -43,7 +43,7 @@ from pipecat.processors.aggregators.openai_llm_context import (
     OpenAILLMContextFrame,
 )
 from pipecat.processors.frame_processor import FrameDirection
-from pipecat.services.ai_services import LLMService
+from pipecat.services.llm_service import LLMService
 
 try:
     from anthropic import NOT_GIVEN, AsyncAnthropic, NotGiven
