+++ src/pipecat/services/azure/stt.py
@@ -16,8 +16,8 @@ from pipecat.frames.frames import (
     StartFrame,
     TranscriptionFrame,
 )
-from pipecat.services.ai_services import STTService
 from pipecat.services.azure.common import language_to_azure_language
+from pipecat.services.stt_service import STTService
 from pipecat.transcriptions.language import Language
 from pipecat.utils.time import time_now_iso8601
