+++ src/pipecat/services/deepgram/tts.py
@@ -16,7 +16,7 @@ from pipecat.frames.frames import (
     TTSStartedFrame,
     TTSStoppedFrame,
 )
-from pipecat.services.ai_services import TTSService
+from pipecat.services.tts_service import TTSService
 
 try:
     from deepgram import DeepgramClient, SpeakOptions
