+++ src/pipecat/services/fal/image.py
@@ -15,7 +15,7 @@ from PIL import Image
 from pydantic import BaseModel
 
 from pipecat.frames.frames import ErrorFrame, Frame, URLImageRawFrame
-from pipecat.services.ai_services import ImageGenService
+from pipecat.services.image_service import ImageGenService
 
 try:
     import fal_client
