+++ src/pipecat/services/gladia/stt.py
@@ -20,8 +20,8 @@ from pipecat.frames.frames import (
     StartFrame,
     TranscriptionFrame,
 )
-from pipecat.services.ai_services import STTService
 from pipecat.services.gladia.config import GladiaInputParams
+from pipecat.services.stt_service import STTService
 from pipecat.transcriptions.language import Language
 from pipecat.utils.time import time_now_iso8601
