+++ src/pipecat/services/openai/tts.py
@@ -17,7 +17,7 @@ from pipecat.frames.frames import (
     TTSStartedFrame,
     TTSStoppedFrame,
 )
-from pipecat.services.ai_services import TTSService
+from pipecat.services.tts_service import TTSService
 
 ValidVoice = Literal["alloy", "echo", "fable", "onyx", "nova", "shimmer"]
