+++ src/pipecat/services/openai_realtime_beta/openai.py
@@ -53,7 +53,7 @@ from pipecat.processors.aggregators.openai_llm_context import (
     OpenAILLMContextFrame,
 )
 from pipecat.processors.frame_processor import FrameDirection
-from pipecat.services.ai_services import LLMService
+from pipecat.services.llm_service import LLMService
 from pipecat.services.openai.llm import OpenAIContextAggregatorPair
 from pipecat.utils.time import time_now_iso8601
