+++ src/pipecat/services/tavus/video.py
@@ -23,7 +23,7 @@ from pipecat.frames.frames import (
     TTSStoppedFrame,
 )
 from pipecat.processors.frame_processor import FrameDirection
-from pipecat.services.ai_services import AIService
+from pipecat.services.ai_service import AIService
 
 
 class TavusVideoService(AIService):
